+++ benchopt/cli/process_results.py
@@ -70,21 +70,22 @@ def plot(benchmark, filename=None, kinds=('suboptimality_curve',),
 )
 @click.argument('benchmark', default=None, type=click.Path(exists=True),
                 required=False, shell_complete=complete_benchmarks)
-@click.option('--filenames', '-f', type=str, multiple=True,
+@click.option('--filename', '-f', 'filenames', type=str, multiple=True,
               shell_complete=complete_output_files,
               help="Specify the files to merge in the benchmark. If it is "
               "not specified, take all files in the benchmark output folder.")
-@click.option('--overwrite', is_flag=True,
-              help="If this flag is set, when merging files, consider that "
-              "they can contain multiple times the same configuration, and "
-              "only keep the last one. This is useful when merging files from "
-              "multiple runs where we add new methods but don't want to lose "
-              "the results of the already existing methods.")
-@click.option('--output', '-o', type=str, default="merged_results.parquet",
+@click.option('--keep', type=click.Choice(['all', 'last']), default='last',
+              help="When merged files contain multiple times the same "
+              "configuration, ``--keep`` controls whether to keep all the "
+              "lines or only keep the last result per configuration. "
+              "This is typically useful when merging files from multiple runs "
+              "where we add new methods but don't want to duplicate the "
+              "results of the already existing methods.")
+@click.option('--output', '-o', type=str, default="merged_results",
               help="Specify the name of the output file. If not specified, "
               "the merged results will be saved in a file named "
               "`merged_results.parquet` in the benchmark output folder.")
-def merge(benchmark, filenames=None, overwrite=False, output=None):
+def merge(benchmark, filenames=None, keep='last', output=None):
 
     if len(filenames) == 0:
         filenames = "all"
@@ -103,10 +104,18 @@ def merge(benchmark, filenames=None, overwrite=False, output=None):
     else:
         output = Path(output)
 
+    # Automatically use the same extension as the input files
+    # if the output filename has the same extensions, otherwise default to
+    # parquet.
+    if output.suffix == "":
+        all_exts = set(f.suffix for f in result_filenames)
+        ext = result_filenames[0].suffix if len(all_exts) == 1 else ".parquet"
+        output = output.with_suffix(ext)
+
     # Merge the results.
     from benchopt.results import save_results
     from benchopt.results.process import merge_results
-    df = merge_results(result_filenames, overwrite=overwrite)
+    df = merge_results(result_filenames, keep=keep)
     save_results(df, output)
 
 
@@ -117,33 +126,56 @@ def merge(benchmark, filenames=None, overwrite=False, output=None):
 )
 @click.argument('benchmark', default=Path.cwd(), type=click.Path(exists=True),
                 shell_complete=complete_benchmarks)
-@click.option('--token', '-t', type=str, default=None,
-              help="Github token to access the result repo.")
 @click.option('--filename', '-f', type=str, default=None,
               shell_complete=complete_output_files,
               help="Specify the file to publish in the benchmark. If it is "
               "not specified, take the latest one in the benchmark output "
               "folder.")
-def publish(benchmark, token=None, filename=None):
-
-    if token is None:
-        token = get_setting('github_token')
-    if token is None:
-        raise RuntimeError(
-            "Could not find the token value to connect to GitHub.\n\n"
-            "Please go to https://github.com/settings/tokens to generate a "
-            "personal token $TOKEN.\nThen, either provide it with option `-t` "
-            "or put it in a config file ./benchopt.yml\nunder section "
-            "[benchopt] as `github_token = $TOKEN`."
-        )
+@click.option('--hub', type=click.Choice(['github', 'huggingface']),
+              default="github", show_default=True,
+              help="Hub to publish the result on. Currently only 'github' and "
+              "'huggingface' are supported.")
+@click.option('--token', '-t', type=str, default=None,
+              help="Github/HF token to access the result repo.")
+@click.option('--repo', '-R', type=str, default=None,
+              help="[HF only] HF repo ID to upload the result to.")
+@click.option('--keep', type=click.Choice(['all', 'last']), default='last',
+              help="[HF only] Results are merged in a single file for hugging "
+              "face. This flag controls whether to keep all lines or only the "
+              "most recent line per unique configuration when merging the new "
+              "results with the existing ones in the repo.")
+def publish(benchmark, filename=None, hub="github", token=None, repo=None,
+            keep='last'):
 
     # Get the result file
     benchmark = Benchmark(benchmark)
     result_filename = benchmark.get_result_files(filename)[0]
 
-    # Publish the result.
-    from benchopt.utils.github import publish_result_file
-    publish_result_file(benchmark, result_filename, token)
+    if hub == "github":
+        if token is None:
+            token = get_setting('github_token')
+        if token is None:
+            raise RuntimeError(
+                "Could not find the token value to connect to GitHub.\n\n"
+                "Please go to https://github.com/settings/tokens to generate "
+                "a personal token $TOKEN.\nThen, either provide it with "
+                "option ``-t``, as an environment variable "
+                "``BENCHOPT_GITHUB_TOKEN``, or put it in a config file "
+                "``./benchopt.yml`` as ``github_token = $TOKEN``."
+            )
+
+        # Publish the result.
+        from benchopt.results.github import publish_result_file
+        publish_result_file(benchmark, result_filename, token)
+    elif hub == "huggingface":
+        if token is None:
+            token = get_setting('hf_token')
+        if repo is None:
+            repo = benchmark.get_setting('hf_repo')
+
+        # Publish the result.
+        from benchopt.results.hugging_face import publish_result_file
+        publish_result_file(benchmark, result_filename, repo, token)
 
 
 @process_results.command(